Add noisy test
